requirements.txt
streamlit
langgraph
langchain
langchain-core
langchain-ollama
python-dotenv
faiss-cpu
